+++ history.md
@@ -1,11 +1,11 @@
-<!--mdast setext-->
+<!--remark setext-->
 
 <!--lint disable no-multiple-toplevel-headings-->
 
 1.0.0 / 2015-08-19
 ==================
 
-*   Update dependencies, dev-dependencies ([feaca83](https://github.com/wooorm/mdast-inline-links/commit/feaca83))
+*   Update dependencies, dev-dependencies ([feaca83](https://github.com/wooorm/remark-inline-links/commit/feaca83))
 
 0.1.0 / 2015-07-12
 ==================